Fixes inflation error
