--- EWRAApp/app/src/main/java/com/kareh/ewraapp/MainActivity.java
@@ -1,14 +1,0 @@
-package com.kareh.ewraapp;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-    }
-}
