+++ EWRAApp/app/src/main/java/com/kareh/ewraapp/ui/HomeActivity.kt
@@ -1,4 +1,4 @@
-package com.kareh.ewraapp
+package com.kareh.ewraapp.ui
 
 import android.os.Bundle
 import com.google.android.material.bottomnavigation.BottomNavigationView
@@ -7,6 +7,7 @@ import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupActionBarWithNavController
 import androidx.navigation.ui.setupWithNavController
+import com.kareh.ewraapp.R
 
 class HomeActivity : AppCompatActivity() {
