+++ EWRAApp/app/src/main/java/com/kareh/ewraapp/ui/MainActivity.java
@@ -1,0 +1,34 @@
+package com.kareh.ewraapp.ui;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Context;
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+
+import com.kareh.ewraapp.R;
+
+public class MainActivity extends AppCompatActivity {
+    private Button btnLogin;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        btnLogin = findViewById(R.id.btn_login);
+        btnLogin.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                dummyLogin(MainActivity.this);
+            }
+        });
+    }
+
+    private void dummyLogin(Context context) {
+        Intent intent = new Intent(context, HomeActivity.class);
+        startActivity(intent);
+    }
+}
