+++ EWRAApp/app/src/main/java/com/kareh/ewraapp/ui/MpesaFragment.java
@@ -1,4 +1,4 @@
-package com.kareh.ewraapp;
+package com.kareh.ewraapp.ui;
 
 import android.os.Bundle;
 
@@ -8,13 +8,15 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.kareh.ewraapp.R;
+
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link ShopFragment#newInstance} factory method to
+ * Use the {@link MpesaFragment#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class ShopFragment extends Fragment {
+public class MpesaFragment extends Fragment {
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
     private static final String ARG_PARAM1 = "param1";
@@ -24,7 +26,7 @@ public class ShopFragment extends Fragment {
     private String mParam1;
     private String mParam2;
 
-    public ShopFragment() {
+    public MpesaFragment() {
         // Required empty public constructor
     }
 
@@ -34,11 +36,11 @@ public class ShopFragment extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment ShopFragment.
+     * @return A new instance of fragment MpesaFragment.
      */
     // TODO: Rename and change types and number of parameters
-    public static ShopFragment newInstance(String param1, String param2) {
-        ShopFragment fragment = new ShopFragment();
+    public static MpesaFragment newInstance(String param1, String param2) {
+        MpesaFragment fragment = new MpesaFragment();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -59,6 +61,6 @@ public class ShopFragment extends Fragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_shop, container, false);
+        return inflater.inflate(R.layout.fragment_mpesa, container, false);
     }
 }
